+++ datacube/index/abstract.py
@@ -1141,11 +1141,14 @@ class AbstractDatasetResource(ABC):
         :param Iterable[Union[str,UUID]] ids: list of dataset ids to archive
         """
 
-    def archive_less_mature(self, ds: Dataset, delta: int = 500) -> None:
+    def archive_less_mature(self, ds: Dataset, delta: Union[int, bool] = 500) -> None:
         """
         Archive less mature versions of a dataset
 
         :param Dataset ds: dataset to search
+        :param Union[int,bool] delta: millisecond delta for time range.
+        If True, default to 500ms. If False, do not find or archive less mature datasets.
+        Bool value accepted only for improving backwards compatibility, int preferred.
         """
         less_mature = self.find_less_mature(ds, delta)
         less_mature_ids = map(lambda x: x.id, less_mature)
@@ -1154,16 +1157,29 @@ class AbstractDatasetResource(ABC):
         for lm_ds in less_mature_ids:
             _LOG.info(f"Archived less mature dataset: {lm_ds}")
 
-    def find_less_mature(self, ds: Dataset, delta: int = 500) -> Iterable[Dataset]:
+    def find_less_mature(self, ds: Dataset, delta: Union[int, bool] = 500) -> Iterable[Dataset]:
         """
         Find less mature versions of a dataset
 
         :param Dataset ds: Dataset to search
-        :param int delta: millisecond delta for time range
+        :param Union[int,bool] delta: millisecond delta for time range.
+        If True, default to 500ms. If None or False, do not find or archive less mature datasets.
+        Bool value accepted only for improving backwards compatibility, int preferred.
         :return: Iterable of less mature datasets
         """
-        less_mature = []
-        assert delta >= 0
+        if isinstance(delta, bool):
+            _LOG.warning("received delta as a boolean value. Int is prefered")
+            if delta is True:  # treat True as default
+                delta = 500
+            else:  # treat False the same as None
+                return []
+        elif isinstance(delta, int):
+            if delta < 0:
+                raise ValueError("timedelta must be a positive integer")
+        elif delta is None:
+            return []
+        else:
+            raise TypeError("timedelta must be None, a positive integer, or a boolean")
 
         def check_maturity_information(dataset, props):
             # check that the dataset metadata includes all maturity-related properties
@@ -1184,6 +1200,7 @@ class AbstractDatasetResource(ABC):
                             region_code=ds.metadata.region_code,
                             time=expanded_time_range)
 
+        less_mature = []
         for dupe in dupes:
             if dupe.id == ds.id:
                 continue