+++ docs/about/whats_new.rst
@@ -34,6 +34,7 @@ v1.8.next
 - Throw a better error if a dataset is not compatible with ``archive_less_mature`` logic (:pull:`1491`)
 - Fix broken Github action workflow (:pull:`1496`)
 - Support ``like=<GeoBox>`` in virtual product ``load`` (:pull:`1497`)
+- Don't archive less mature if archive_less_mature is provided as `False` instead of `None` (:pull:`1498`)
 
 v1.8.15 (11th July 2023)
 ========================